Extract ChannelStoreEntry interface
