+++ packages/xstate-wallet/src/store/channel-store-entry.ts
@@ -1,0 +1,12 @@
+import {ChannelConstants, StateVariables, State} from './types';
+import {Funding} from './memory-store';
+export interface ChannelStoreEntry {
+  readonly channelId: string;
+  readonly myIndex: number;
+  readonly latest: StateVariables;
+  readonly supported: StateVariables | undefined;
+  readonly latestSupportedByMe: StateVariables | undefined;
+  readonly channelConstants: ChannelConstants;
+  readonly funding?: Funding;
+  readonly states: State[];
+}
